add mooncrater-input implementation
This was actually many commits over a long time, but I leaned heavily on AI generating code, and I didn't want to keep so much intermediate generated garbage.
I tried using a style of giving to-do lists, and got tons of intermediate stuff committed, including so many wrong directions that I fixed forward with the to-do list workflow, and it would have been a pain to clean to a nice and useful history.
So I decided to just squash it all now that I've finally “finished” it to where I want to use it as my main implementation of my keyboard layout on Linux.